vgg_service/requirements.txt
Flask==2.1.2
librosa==0.8.1
numpy==1.19.5
tensorflow==2.6.0
pandas==1.3.5  
matplotlib==3.4.3 
Pillow==8.2.0